requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.0.3
flask-jwt-extended==4.5.3
flask-socketio==5.3.6
python-dotenv==1.0.0
gunicorn==21.2.0
web3==6.15.1
tonclient-py==2.0.0
requests==2.31.0
python-telegram-bot==20.6
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
